feat: import geopackage, move create-search-index into postgres/load package
